Soldier can now be dragged around and attacking has begun implementation
Test out creating a new private variable in GameManager, that specifies whether the player is supposed to raise onmove or not, which is set to false when the player right-clicks on the enemy. After that, the player sets the variable to true again since that should be the default value.