Swordsman now attacks and hurts enemies! Amazing!
There is a weird little one jump the swordsman does after killing an enemy, fix that even though its cute. Also change isIdle variable to, doNotDisturb and make that have full priority in OnTriggerEnter2D and set it to true when soldier is attacking, set it to false when moving in attacking coroutine. That way stop soldiers from overlapping when attacking enemies